requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
python-dotenv
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
networkx
pyvis
httpx
pillow
watchdog
click
# Force Rebuild 1